# docs/openapi.yaml
openapi: 3.0.1
info:
  title: go-transactions
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: localhost
paths:
  /accounts:
    post:
      tags:
        - accounts
      description: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Account"
      responses:
        201:
          description: Account created with success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        500:
          description: Internal server error
  /accounts/{accountId}:
    get:
      tags:
        - accounts
      description: Get account informations
      parameters:
        - in: path
          name: accountId
          description: The account identifier
          required: true
          schema:
            type: number
      responses:
        200:
          description: Account got with success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        500:
          description: Internal server error
  /transactions:
    post:
      tags:
        - transactions
      description: Register a new transaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Transaction"
      responses:
        201:
          description: Transaction created with successful
        400:
          description: Bad request
        500:
          description: Internal server error
    get:
      tags:
        - transactions
      description: Get all transactions
      responses:
        200:
          description: Account got with success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        500:
          description: Internal server error
components:
  schemas:
    Account:
      type: object
      properties:
        id:
          type: number
          example: 1
        document_number:
          type: string
          example: "12345678900"
        available_credit_limit:
          type: number
          example: 1000.00
    Transaction:
      type: object
      properties:
        account_id:
          type: number
          example: 1
        operation_type_id:
          type: number
          example: 4
        amount:
          type: number
          example: 123.45